requirements.txt
# AI Model providers - choose one or both
openai==1.40.0  # For OpenAI API (optional)
ollama>=0.3.0  # For local Ollama models
langchain==0.2.14
langchain-openai==0.1.20  # For OpenAI support (optional)
langchain-ollama>=0.1.0  # For Ollama support
langchain-community==0.2.12
langchain-core>=0.2.43
langchain-text-splitters>=0.2.4
speechrecognition==3.10.4
pyaudio==0.2.14
pydub==0.25.1
streamlit==1.37.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
seaborn==0.13.2
requests==2.32.3

# Jupyter notebook support
ipykernel>=6.28.0
jupyter>=1.0.0
ipython>=8.20.0
